+++ max_div/solver/_diversity/_base.py
@@ -5,7 +5,13 @@ from typing import Callable
 import numpy as np
 from numpy.typing import NDArray
 
-from ._numba import approx_geomean_separation, geomean_separation, mean_separation, min_separation
+from ._numba import (
+    approx_geomean_separation,
+    geomean_separation,
+    mean_separation,
+    min_separation,
+    non_zero_separation_frac,
+)
 
 
 class DiversityMetric:
@@ -17,6 +23,7 @@ class DiversityMetric:
         geomean_separation()         Geometric mean separation of all selected vectors
         approx_geomean_separation()  Approximate geometric mean separation of all selected vectors
                                          (uses faster, but still smooth approximations of log(.) and exp(.))
+        non_zero_separation_frac()   Fraction of separation values that are non-zero
     """
 
     # -------------------------------------------------------------------------
@@ -32,6 +39,9 @@ class DiversityMetric:
             return np.float32(0.0)
         else:
             return self.f(separations)
+
+    def __eq__(self, other) -> bool:
+        return isinstance(other, DiversityMetric) and (self.name == other.name) and (self.f == other.f)
 
     # -------------------------------------------------------------------------
     #  Factory methods
@@ -55,3 +65,8 @@ class DiversityMetric:
     def approx_geomean_separation(cls) -> DiversityMetric:
         """Approximate geometric mean separation of all selected vectors."""
         return cls(name="approx_geomean_separation", f=approx_geomean_separation)
+
+    @classmethod
+    def non_zero_separation_frac(cls) -> DiversityMetric:
+        """Fraction of separation values that are non-zero."""
+        return cls(name="non_zero_separation_frac", f=non_zero_separation_frac)