+++ tests/solver/_diversity/test_base.py
@@ -5,7 +5,7 @@ from max_div.solver._diversity import DiversityMetric
 
 
 @pytest.mark.parametrize(
-    "metric, separation, expected_result, rel_tol",
+    "metric, separation, expected_result, tol",
     [
         (DiversityMetric.min_separation(), [], 0.0, 1e-6),
         (DiversityMetric.mean_separation(), [], 0.0, 1e-6),
@@ -17,9 +17,12 @@ from max_div.solver._diversity import DiversityMetric
         (DiversityMetric.geomean_separation(), [0.1, 0.0], 0.0, 1e-6),
         (DiversityMetric.approx_geomean_separation(), [0.1, 0.4], 0.2, 0.01),
         (DiversityMetric.approx_geomean_separation(), [0.1, 0.0], 0.0, 1e-6),
+        (DiversityMetric.non_zero_separation_frac(), [0.1, 0.4], 1.0, 1e-6),
+        (DiversityMetric.non_zero_separation_frac(), [0.1, 0.0], 0.5, 1e-6),
+        (DiversityMetric.non_zero_separation_frac(), [0.0, 0.0], 0.0, 1e-6),
     ],
 )
-def test_diversity_compute(metric: DiversityMetric, separation: list[float], expected_result: float, rel_tol: float):
+def test_diversity_compute(metric: DiversityMetric, separation: list[float], expected_result: float, tol: float):
     # --- arrange -----------------------------------------
     separation = np.array(separation, dtype=np.float32)
 
@@ -27,4 +30,13 @@ def test_diversity_compute(metric: DiversityMetric, separation: list[float], exp
     result = metric.compute(separation)
 
     # --- assert ------------------------------------------
-    assert result == pytest.approx(expected_result, rel=rel_tol)
+    assert result == pytest.approx(expected_result, abs=tol, rel=tol)
+
+
+def test_diversity_metric_equals():
+    metric1 = DiversityMetric.min_separation()
+    metric2 = DiversityMetric.min_separation()
+    metric3 = DiversityMetric.mean_separation()
+
+    assert metric1 == metric2
+    assert metric1 != metric3